Filename change & button feedback removed.
Filename changed to reflect its type, with ino_Proton.ino being the main
sketch.

Button feedback removed to prevent spamming the particle cloud event
log. (lights turn on and off multiple times a day)
