+++ lib_PietteTech_DHT.cpp
@@ -42,7 +42,7 @@
 
 #include "application.h"
 #include "math.h"
-#include "PietteTech_DHT.h"
+#include "lib_PietteTech_DHT.h"
 
 // Thanks to Paul Kourany for this word type conversion function
 uint16_t word(uint8_t high, uint8_t low) {
